+++ src/localops/tools/__init__.py
@@ -1,5 +1,6 @@
 """Predefined read-only server tools."""
 
+from .cpu import get_cpu_load
 from .disk import get_disk_usage
 from .errors import ToolCommandError
 from .memory import get_memory_usage
@@ -7,6 +8,7 @@ from .system import get_system_info
 
 __all__ = [
     "ToolCommandError",
+    "get_cpu_load",
     "get_disk_usage",
     "get_memory_usage",
     "get_system_info",